removed more unused functions from run_pyrate
